+++ common/src/main/java/net/mehvahdjukaar/amendments/integration/CompatHandler.java
@@ -25,7 +25,7 @@ public class CompatHandler {
     public static final boolean ALEX_CAVES = PlatHelper.isModLoaded("alexscaves");
     public static final boolean RATS = PlatHelper.isModLoaded("rats");
     public static final boolean THIN_AIR = PlatHelper.isModLoaded("thinair");
-    public static final boolean FLYWHEEL = PlatHelper.isModLoaded("flywheel") && false;
+    public static final boolean VANILLIN = false;// PlatHelper.isModLoaded("vanillin") && false;
 
     static {
         boolean fd = false;